Jiminy 1.7.4 (#444)
* [core] Implement friction cone for impulse model. (#440)
* [core] Fix numerical instability of flexibility model by adding armature-like inertia parameter. (#433)
* [core] Full support of armature for all joints but mimic.
* [core] Rename 'pncCollisionData' in 'collisionData' and 'pncModelRigidOrig_' in 'pncModelOrig_' since already clear.
* [core] Fix some implicit type conversions.
* [python/robot] Fix default hardware config file generator is both collision and visual meshes are defined.
* [python/viewer] Add physic based rendering of materials to panda3d. 
* [python/viewer] Fix shadow casting area smaller than floor in panda3d. 
* [python/viewer] Fix window name not set properly in panda3d.
* [python/viewer] Change direction of lighting in panda3d to fell natural for default camera pose.
* [python/viewer] Enhanced rendering in panda3d for discrete gpu.
* [python/viewer] Render meshes two-sided in panda3d to avoid seeing through them.
* [gym/rllib] Add option to forward keyword arguments to 'ray.init'. (#436)
* [misc] Update Python dependencies.
* [misc] Update Boost from 1.7.1 to 1.7.6.
* [misc] Build OSX wheels. (#442)

+++ core/src/constraints/AbstractConstraint.cc
@@ -96,7 +96,7 @@ namespace jiminy
 
     uint64_t AbstractConstraintBase::getDim(void) const
     {
-        return drift_.size();
+        return static_cast<uint64_t>(drift_.size());
     }
 
     matrixN_t const & AbstractConstraintBase::getJacobian(void) const
